requirements.txt
streamlit>=1.33.0
tensorflow>=2.10.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
